Use keystores directly until we figure out how to use exported certs.
What is missing in the manually constructed key stores is the private keys
corresponding to the imported certificates, as well as the CA certs.

I've tried importing PKCS12 files (.p12) because they actually contain the
private key along with the rest of the cert, but to no avail yet. The idea
was to load "trusts-ca.jks" as a "prototype", as it already trusts the CA,
then add the .p12 certs that were discovered.